Allocate memory from arena with correct alignment
The arena allocator does not align memory when allocating blocks which can
trigger undefined behaviour. From UBSAN:

  src/inlines.c:83:30: runtime error: member access within misaligned address
  0x00010b30ede3 for type 'cmark_node' (aka 'struct cmark_node'), which requires
  8 byte alignment

This could cause a crash if the unaligned memory is passed to functions using
instructions which require alignment, such as SSE. Round-up the size of the
allocation to a multiple of `sizeof(size_t)` to ensure every allocation is
correctly aligned.